backend/pyproject.toml

[project]
name = "storage_manager"
version = "0.2.0"
requires-python = ">=3.9"
dependencies = [
    "fastapi", 
    "uvicorn[standard]", 
    "sqlalchemy", 
    "pydantic", 
    "python-dotenv", 
    "starlette>=0.46.1",
    "alembic>=1.15.2",
    "pylabels>=1.2.1",
    "reportlab>=4.4.1",
    "jinja2>=3.1.6",
]

[dependency-groups]
dev = [
    "pytest",
    "flake8",
    "black",
    "pre-commit>=4.2.0",
    "httpx",
    "pytest",
    "flake8",
    "black",
    "pre-commit",
    "deepdiff>=8.5.0",
    "pypdf>=5.6.0",
]

[tool.pdm.scripts]
migrate = "alembic upgrade head"
migrationtest = "alembic/migration_test.sh"
makemigration = "alembic revision --autogenerate -m"
init_db = "python -m app.scripts.init_db"

[tool.pdm]
package = [{include = "app"},   { include = "alembic" }]

[tool.setuptools]
packages = ["app", "alembic"]

[tool.pytest.ini_options]
pythonpath = ["."]
norecursedirs = ["__pypackages__"]